fix: TeamService 구현체 분리
